move construct-based parser to construct_mp4_parser.py
