+++ tests/cli/simple_mp4_parser.py
@@ -7,6 +7,7 @@ import typing as T
 
 from mapillary_tools import utils
 from mapillary_tools.geotag import (
+    construct_mp4_parser as cparser,
     mp4_sample_parser as sample_parser,
     simple_mp4_builder as builder,
     simple_mp4_parser as parser,
@@ -105,7 +106,7 @@ def _parse_samples(fp: T.BinaryIO, filters: T.Optional[T.Container[bytes]] = Non
     for h, s in parser.parse_path(fp, [b"moov", b"trak"]):
         offset = s.tell()
         for h1, s1 in parser.parse_path(s, [b"mdia", b"mdhd"], maxsize=h.maxsize):
-            box = parser.MediaHeaderBox.parse(s1.read(h.maxsize))
+            box = cparser.MediaHeaderBox.parse(s1.read(h.maxsize))
             LOG.info(box)
             LOG.info(sample_parser.to_datetime(box.creation_time))
             LOG.info(box.duration / box.timescale)
@@ -206,10 +207,10 @@ def _process_path(parsed_args, path: pathlib.Path):
                         _parse_structs(io.BytesIO(data))
                 elif parsed_args.full:
                     if box_path is None:
-                        boxes = parser.FullBoxStruct64.BoxList.parse_stream(fp)
+                        boxes = cparser.FullBoxStruct64.BoxList.parse_stream(fp)
                     else:
                         data = parser.parse_mp4_data_firstx(fp, box_path)
-                        boxes = parser.FullBoxStruct64.BoxList.parse_stream(
+                        boxes = cparser.FullBoxStruct64.BoxList.parse_stream(
                             io.BytesIO(data)
                         )
                     print(boxes)
