+++ tests/unit/test_simple_mp4_builder.py
@@ -1,10 +1,12 @@
 import io
 import typing as T
 
-import mapillary_tools.geotag.mp4_sample_parser as sample_parser
-
-import mapillary_tools.geotag.simple_mp4_builder as builder
-import mapillary_tools.geotag.simple_mp4_parser as parser
+from mapillary_tools.geotag import (
+    construct_mp4_parser as cparser,
+    mp4_sample_parser as sample_parser,
+    simple_mp4_builder as builder,
+    simple_mp4_parser as parser,
+)
 
 
 def _test_build(b, path):
@@ -20,20 +22,20 @@ def test_build_sample_5s():
     simple_mp4 = (
         "tests/integration/mapillary_tools_process_images_provider/data/sample-5s.mp4"
     )
-    _test_build(parser.FullBoxStruct32.BoxList, simple_mp4)
-    _test_build(parser.QuickBoxStruct32.BoxList, simple_mp4)
-    _test_build(parser.FullBoxStruct64.BoxList, simple_mp4)
-    _test_build(parser.QuickBoxStruct64.BoxList, simple_mp4)
+    _test_build(cparser.FullBoxStruct32.BoxList, simple_mp4)
+    _test_build(cparser.QuickBoxStruct32.BoxList, simple_mp4)
+    _test_build(cparser.FullBoxStruct64.BoxList, simple_mp4)
+    _test_build(cparser.QuickBoxStruct64.BoxList, simple_mp4)
 
 
 def test_build_hero():
     hero_mp4 = (
         "tests/integration/mapillary_tools_process_images_provider/gopro_data/hero8.mp4"
     )
-    _test_build(parser.FullBoxStruct32.BoxList, hero_mp4)
-    _test_build(parser.QuickBoxStruct32.BoxList, hero_mp4)
-    _test_build(parser.FullBoxStruct64.BoxList, hero_mp4)
-    _test_build(parser.QuickBoxStruct64.BoxList, hero_mp4)
+    _test_build(cparser.FullBoxStruct32.BoxList, hero_mp4)
+    _test_build(cparser.QuickBoxStruct32.BoxList, hero_mp4)
+    _test_build(cparser.FullBoxStruct64.BoxList, hero_mp4)
+    _test_build(cparser.QuickBoxStruct64.BoxList, hero_mp4)
 
 
 def _build_and_parse_stbl(
@@ -43,7 +45,7 @@ def _build_and_parse_stbl(
         descriptions,
         expected_samples,
     )
-    d = parser.FullBoxStruct32.Box.build({"type": b"stbl", "data": s})
+    d = cparser.FullBoxStruct32.Box.build({"type": b"stbl", "data": s})
     ss = parser.parse_box_data_firstx(io.BytesIO(d), [b"stbl"])
     assert d[8:] == ss
     _, parsed_samples = sample_parser.parse_raw_samples_from_stbl(io.BytesIO(ss))
@@ -120,7 +122,7 @@ def test_parse_raw_samples_from_stbl():
     # chunk 2
     #     sample 3: offset=5, size=3, timedelta=30, is_sync=True
     #     sample 4: offset=8, size=3, timedelta=50, is_sync=False
-    stbl_bytes = parser.FullBoxStruct32.BoxList.build(
+    stbl_bytes = cparser.FullBoxStruct32.BoxList.build(
         [
             {
                 "type": b"stsd",
@@ -247,36 +249,36 @@ def test_parse_raw_samples_from_stbl():
 
 
 def test_box_header_0_building():
-    data = parser.BoxHeader0.build(
+    data = cparser.BoxHeader0.build(
         {
             "type": b"ftyp",
         }
     )
     assert data == b"\x00\x00\x00\x00ftyp"
-    p = parser.BoxHeader0.parse(data)
+    p = cparser.BoxHeader0.parse(data)
     assert p["size32"] == 0
 
 
 def test_box_header_32_building():
-    data = parser.BoxHeader32.build(
+    data = cparser.BoxHeader32.build(
         {
             "size": 123,
             "type": b"ftyp",
         }
     )
     assert data == b"\x00\x00\x00{ftyp"
-    p = parser.BoxHeader32.parse(data)
+    p = cparser.BoxHeader32.parse(data)
     assert p["size"] == 123
 
 
 def test_box_header_64_building():
-    data = parser.BoxHeader64.build(
+    data = cparser.BoxHeader64.build(
         {
             "size": 123,
             "type": b"ftyp",
         }
     )
     assert data == b"\x00\x00\x00\x01ftyp\x00\x00\x00\x00\x00\x00\x00{"
-    p = parser.BoxHeader64.parse(data)
+    p = cparser.BoxHeader64.parse(data)
     assert p["size"] == 123
     assert p["size32"] == 1
